Address comments from @Francois-Esquire
